requirements.txt
jupyterlab
jupyter_http_over_ws
numpy
pandas
matplotlib
lang2vec
umap-learn
ipympl
pycountry
basemap